requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
google-generativeai
streamlit
asyncio
pyngrok
sentence-transformers
torch